Improve variance computation, simplify while loop conditions, use sort built-in function
The variances of the estimations of the shapley values are now computed using the maximum likelihood estimator -- not the moment estimator as before. It now uses a little more memory but it has the advantage of always yielding a positive estimation of the variance.

This commit removed the condition to ensure the first 50 permutations  were different in the while loop of the truncated_MC method. Have the  50 permutations equal is an extremely unlikely event, this condition was therefore unnecessary.