Paint in black + add TMC2
TMC2 is a small modification of the TMC method, its goal is to reduce the bias. In TMC2, instead of doing a truncation to zero of the last increments of the permutation,  we do a linear interpolation. This should be a better approximation of the last increments of the permutation, hence it should reduce the bias of the method.