Fix smodel initialization Fix kwargs use for mpl in contributivity
S-model initialization could fail if the confusion matrix has not the right shape, which can be the case if some labels are not included in the dataset of a partner. btw I noticed that smodel can only work with datasets with 10 labels, that's only cifar and mnist. I opened an issue about that
Signed-off-by: arthurPignet <arthur.pignet@mines-paristech.fr>